[refactor-file-names] Refactor bundle landing component's name
